move helpers out of controllers
