+++ controller/linodemachine_controller.go
@@ -28,7 +28,6 @@ import (
 	"github.com/linode/linodego"
 	corev1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
-	"k8s.io/apimachinery/pkg/api/resource"
 	utilerrors "k8s.io/apimachinery/pkg/util/errors"
 	"k8s.io/client-go/tools/record"
 	clusterv1 "sigs.k8s.io/cluster-api/api/v1beta1"
@@ -66,12 +65,6 @@ const (
 	ConditionPreflightBootTriggered          clusterv1.ConditionType = "PreflightBootTriggered"
 	ConditionPreflightReady                  clusterv1.ConditionType = "PreflightReady"
 )
-
-var skippedMachinePhases = map[string]bool{
-	string(clusterv1.MachinePhasePending): true,
-	string(clusterv1.MachinePhaseFailed):  true,
-	string(clusterv1.MachinePhaseUnknown): true,
-}
 
 // statuses to keep requeueing on while an instance is booting
 var requeueInstanceStatuses = map[linodego.InstanceStatus]bool{
@@ -125,16 +118,29 @@ func (r *LinodeMachineReconciler) Reconcile(ctx context.Context, req ctrl.Reques
 		return ctrl.Result{}, err
 	}
 
-	machine, err := getOwnerMachine(ctx, r.TracedClient(), *linodeMachine, log)
+	machine, err := kutil.GetOwnerMachine(ctx, r.TracedClient(), linodeMachine.ObjectMeta)
 	if err != nil || machine == nil {
 		return ctrl.Result{}, err
 	}
 	log = log.WithValues("LinodeMachine", machine.Name)
 
-	cluster, err := r.getClusterFromMetadata(ctx, *machine, log)
+	cluster, err := kutil.GetClusterFromMetadata(ctx, r.TracedClient(), machine.ObjectMeta)
 	if err != nil || cluster == nil {
 		return ctrl.Result{}, err
 	}
+
+	linodeClusterKey := client.ObjectKey{
+		Namespace: linodeMachine.Namespace,
+		Name:      cluster.Spec.InfrastructureRef.Name,
+	}
+	linodeCluster := &infrav1alpha2.LinodeCluster{}
+	if err := r.Client.Get(ctx, linodeClusterKey, linodeCluster); err != nil {
+		if err = client.IgnoreNotFound(err); err != nil {
+			return ctrl.Result{}, fmt.Errorf("get linodecluster %q: %w", linodeClusterKey, err)
+		}
+	}
+
+	log = log.WithValues("LinodeCluster", linodeCluster.Name)
 
 	machineScope, err := scope.NewMachineScope(
 		ctx,
@@ -143,7 +149,7 @@ func (r *LinodeMachineReconciler) Reconcile(ctx context.Context, req ctrl.Reques
 			Client:        r.TracedClient(),
 			Cluster:       cluster,
 			Machine:       machine,
-			LinodeCluster: &infrav1alpha2.LinodeCluster{},
+			LinodeCluster: linodeCluster,
 			LinodeMachine: linodeMachine,
 		},
 	)
@@ -256,16 +262,6 @@ func (r *LinodeMachineReconciler) reconcile(
 	res, err = r.reconcileCreate(ctx, logger, machineScope)
 
 	return
-}
-
-func retryIfTransient(err error) (ctrl.Result, error) {
-	if util.IsRetryableError(err) {
-		if linodego.ErrHasStatus(err, http.StatusTooManyRequests) {
-			return ctrl.Result{RequeueAfter: reconciler.DefaultLinodeTooManyRequestsErrorRetryDelay}, nil
-		}
-		return ctrl.Result{RequeueAfter: reconciler.DefaultMachineControllerRetryDelay}, nil
-	}
-	return ctrl.Result{}, err
 }
 
 func (r *LinodeMachineReconciler) reconcileCreate(
@@ -422,174 +418,6 @@ func (r *LinodeMachineReconciler) reconcileInstanceCreate(
 	machineScope.LinodeMachine.Status.InstanceState = util.Pointer(linodego.InstanceOffline)
 
 	return ctrl.Result{}, nil
-}
-
-func (r *LinodeMachineReconciler) configureDisks(
-	ctx context.Context,
-	logger logr.Logger,
-	machineScope *scope.MachineScope,
-	linodeInstanceID int,
-) error {
-	if machineScope.LinodeMachine.Spec.DataDisks == nil && machineScope.LinodeMachine.Spec.OSDisk == nil {
-		return nil
-	}
-
-	if err := r.resizeRootDisk(ctx, logger, machineScope, linodeInstanceID); err != nil {
-		return err
-	}
-	if !reconciler.ConditionTrue(machineScope.LinodeMachine, ConditionPreflightAdditionalDisksCreated) {
-		if err := r.createDisks(ctx, logger, machineScope, linodeInstanceID); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func (r *LinodeMachineReconciler) createDisks(ctx context.Context, logger logr.Logger, machineScope *scope.MachineScope, linodeInstanceID int) error {
-	for deviceName, disk := range machineScope.LinodeMachine.Spec.DataDisks {
-		if disk.DiskID != 0 {
-			continue
-		}
-		label := disk.Label
-		if label == "" {
-			label = deviceName
-		}
-		// create the disk
-		diskFilesystem := defaultDiskFilesystem
-		if disk.Filesystem != "" {
-			diskFilesystem = disk.Filesystem
-		}
-		linodeDisk, err := machineScope.LinodeClient.CreateInstanceDisk(
-			ctx,
-			linodeInstanceID,
-			linodego.InstanceDiskCreateOptions{
-				Label:      label,
-				Size:       int(disk.Size.ScaledValue(resource.Mega)),
-				Filesystem: diskFilesystem,
-			},
-		)
-		if err != nil {
-			if !linodego.ErrHasStatus(err, linodeBusyCode) {
-				logger.Error(err, "Failed to create disk", "DiskLabel", label)
-			}
-
-			conditions.MarkFalse(
-				machineScope.LinodeMachine,
-				ConditionPreflightAdditionalDisksCreated,
-				string(cerrs.CreateMachineError),
-				clusterv1.ConditionSeverityWarning,
-				err.Error(),
-			)
-			return err
-		}
-		disk.DiskID = linodeDisk.ID
-		machineScope.LinodeMachine.Spec.DataDisks[deviceName] = disk
-	}
-	err := r.UpdateInstanceConfigProfile(ctx, logger, machineScope, linodeInstanceID)
-	if err != nil {
-		return err
-	}
-	conditions.MarkTrue(machineScope.LinodeMachine, ConditionPreflightAdditionalDisksCreated)
-	return nil
-}
-
-func (r *LinodeMachineReconciler) resizeRootDisk(
-	ctx context.Context,
-	logger logr.Logger,
-	machineScope *scope.MachineScope,
-	linodeInstanceID int,
-) error {
-	if reconciler.ConditionTrue(machineScope.LinodeMachine, ConditionPreflightRootDiskResized) {
-		return nil
-	}
-
-	instanceConfig, err := r.getDefaultInstanceConfig(ctx, machineScope, linodeInstanceID)
-	if err != nil {
-		logger.Error(err, "Failed to get default instance configuration")
-
-		conditions.MarkFalse(machineScope.LinodeMachine, ConditionPreflightRootDiskResized, string(cerrs.CreateMachineError), clusterv1.ConditionSeverityWarning, err.Error())
-		return err
-	}
-
-	if instanceConfig.Devices.SDA == nil {
-		conditions.MarkFalse(machineScope.LinodeMachine, ConditionPreflightRootDiskResized, string(cerrs.CreateMachineError), clusterv1.ConditionSeverityWarning, "root disk not yet ready")
-
-		return errors.New("root disk not yet ready")
-	}
-
-	rootDiskID := instanceConfig.Devices.SDA.DiskID
-
-	// carve out space for the etcd disk
-	if !reconciler.ConditionTrue(machineScope.LinodeMachine, ConditionPreflightRootDiskResizing) {
-		rootDisk, err := machineScope.LinodeClient.GetInstanceDisk(ctx, linodeInstanceID, rootDiskID)
-		if err != nil {
-			logger.Error(err, "Failed to get root disk for instance")
-
-			conditions.MarkFalse(machineScope.LinodeMachine, ConditionPreflightRootDiskResizing, string(cerrs.CreateMachineError), clusterv1.ConditionSeverityWarning, err.Error())
-
-			return err
-		}
-		// dynamically calculate root disk size unless an explicit OS disk is being set
-		additionalDiskSize := 0
-		for _, disk := range machineScope.LinodeMachine.Spec.DataDisks {
-			additionalDiskSize += int(disk.Size.ScaledValue(resource.Mega))
-		}
-		diskSize := rootDisk.Size - additionalDiskSize
-		if machineScope.LinodeMachine.Spec.OSDisk != nil {
-			diskSize = int(machineScope.LinodeMachine.Spec.OSDisk.Size.ScaledValue(resource.Mega))
-		}
-
-		if err := r.ResizeDisk(ctx, logger, machineScope, linodeInstanceID, rootDiskID, diskSize); err != nil {
-			return err
-		}
-
-		conditions.MarkTrue(machineScope.LinodeMachine, ConditionPreflightRootDiskResizing)
-	}
-
-	conditions.Delete(machineScope.LinodeMachine, ConditionPreflightRootDiskResizing)
-	conditions.MarkTrue(machineScope.LinodeMachine, ConditionPreflightRootDiskResized)
-
-	return nil
-}
-
-func (r *LinodeMachineReconciler) ResizeDisk(ctx context.Context, logger logr.Logger, machineScope *scope.MachineScope, linodeInstanceID, rootDiskID, diskSize int) error {
-	if err := machineScope.LinodeClient.ResizeInstanceDisk(ctx, linodeInstanceID, rootDiskID, diskSize); err != nil {
-		if !linodego.ErrHasStatus(err, linodeBusyCode) {
-			logger.Error(err, "Failed to resize root disk")
-		}
-
-		conditions.MarkFalse(machineScope.LinodeMachine, ConditionPreflightRootDiskResizing, string(cerrs.CreateMachineError), clusterv1.ConditionSeverityWarning, err.Error())
-
-		return err
-	}
-	return nil
-}
-
-func (r *LinodeMachineReconciler) UpdateInstanceConfigProfile(
-	ctx context.Context,
-	logger logr.Logger,
-	machineScope *scope.MachineScope,
-	linodeInstanceID int,
-) error {
-	// get the default instance config
-	configs, err := machineScope.LinodeClient.ListInstanceConfigs(ctx, linodeInstanceID, &linodego.ListOptions{})
-	if err != nil || len(configs) == 0 {
-		logger.Error(err, "Failed to list instance configs")
-
-		return err
-	}
-	instanceConfig := configs[0]
-
-	if machineScope.LinodeMachine.Spec.DataDisks != nil {
-		if err := createInstanceConfigDeviceMap(machineScope.LinodeMachine.Spec.DataDisks, instanceConfig.Devices); err != nil {
-			return err
-		}
-	}
-	if _, err := machineScope.LinodeClient.UpdateInstanceConfig(ctx, linodeInstanceID, instanceConfig.ID, linodego.InstanceConfigUpdateOptions{Devices: instanceConfig.Devices}); err != nil {
-		return err
-	}
-
-	return nil
 }
 
 func (r *LinodeMachineReconciler) reconcileUpdate(
